app/requirements.txt
flask
psycopg2-binary
pandas
sqlalchemy